requirements.txt
fastapi
uvicorn
python-dotenv
supabase
fuzzywuzzy
python-Levenshtein
